requirements.txt
amqp==2.5.2
astroid==2.3.3
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.2
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-jsonapi==3.0.0
dnspython==1.16.0
eventlet==0.25.2
greenlet==0.4.16
gunicorn==20.0.4
html5lib==1.1
idna==2.9
inflection==0.3.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.8
lazy-object-proxy==1.4.3
localStoragePy==0.1.7
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
numpy==1.18.1
pandas==1.0.1
paramiko==2.7.1
Pillow==7.2.0
psycopg2==2.8.4
pycparser==2.19
pylint==2.4.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
reportlab==3.5.49
requests==2.23.0
scp==0.13.2
simplejson==3.17.0
six==1.14.0
SQLAlchemy==1.3.20
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.11.2
xhtml2pdf==0.2.4
xmljson==0.2.1
zope.event==4.5.0
zope.interface==5.1.2
